new economy commands and cleanup
added a beg and balance command  as well as adding a config file at ./settings.json and removed the previous config.json. I also added a beg command which gives users Bottlecaps which is the currency and added a bal command which shows the user's balance. Currently the balance command has issues and will not show a balance but an undefined amount.